Add Super User. CRUD

+++ database/migrations/2025_09_02_020338_add_role_to_users_table.php
@@ -9,15 +9,19 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->string('role')->default('apotek')->after('email'); 
-            // default apotek, bisa 'busdev' juga
+            if (!Schema::hasColumn('users', 'role')) {
+                $table->string('role')->default('apotek')->after('email');
+                // roles: super, busdev, apotek
+            }
         });
     }
 
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('role');
+            if (Schema::hasColumn('users', 'role')) {
+                $table->dropColumn('role');
+            }
         });
     }
 };
